requirements.txt
# Dependencies already installed at GitHub Autograding
black<=22.3.0
pytest~=7.2.2
pytest-subtests~=0.11.0
tomli>=1.1.0; python_full_version < '3.11.2'

# Use CPU-only PyTorch wheels
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

# Common packages for inference scripts
numpy==2.2.6
tqdm==4.67.1
pillow==11.2.1
pyparsing==3.2.3
matplotlib==3.10.3
statistics==1.0.3.5
Path==17.1.0
image==1.5.33
fonttools==4.58.1
Django==5.2.1
docutils==0.21.2
six==1.17.0
sqlparse==0.5.3
ddt==1.7.2
parameterized==0.9.0
